refactor: :recycle: Refactor the cleanup and config
