+++ src/agent/__init__.py
@@ -7,5 +7,4 @@ __app_name__ = "agent"
 __version__ = importlib.metadata.version("automators-agent")
 
 dotenv_path = Path(os.getcwd()) / ".env"
-
 load_dotenv(dotenv_path=dotenv_path, override=True)
