+++ src/agent/config.py
@@ -1,0 +1,4 @@
+from pathlib import Path
+
+# create a folder to store the generated code
+TEST_DIR = Path("tests")
