+++ src/agent/tools.py
@@ -1,6 +1,6 @@
 import os
 from playwright.sync_api import sync_playwright
-from agent.setup import TEST_DIR
+from agent.config import TEST_DIR
 from agent.utils import run_pytest_and_capture_output, strip_code_fences
 from agent.logging import logger, console
 from rich.panel import Panel
@@ -62,8 +62,12 @@ def write_code_to_file(**kwargs: TWriteCodeToFile):
 
 
 def run_tests():
+    # ensure the tests directory exists
+    TEST_DIR.mkdir(exist_ok=True)
     logger.info("Running tests")
+    # run the tests and capture the output
     output = run_pytest_and_capture_output(TEST_DIR)
+    # print output in a panel
     console.print("\n")
     console.print(Panel(output, title="Test Output", highlight=True, padding=(1, 1)))
     console.print("\n")
